requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==23.2.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/ultralytics/CLIP.git@4f352ca0a7450598833a770314f73e1f74e05f30
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
exceptiongroup==1.3.1
fastapi==0.123.0
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.10.0
ftfy==6.3.1
gradio==3.50.2
gradio_client==0.6.1
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.13.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==10.4.0
pillow_heif==1.1.1
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
protobuf==6.33.1
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
pyglet==1.5.31
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
roma==1.5.4
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.1+cu126
torchvision==0.24.1+cu126
tqdm==4.67.1
transformers==4.57.3
trimesh==4.10.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==11.0.3
Werkzeug==3.1.4
